[parser][sandbox] Validate Code::kMaxArguments for wrapped functions
An in-sandbox attacker can corrupt a lazy function's SFI flags to
kWrapped and attach a 65,536-element wrapped_arguments FixedArray. The
parser does not validate the wrapped function arguments against
Code::kMaxArguments. During bytecode compilation, the parameter_count
overflows int -> uint16_t (65,537 to 1). This results in stack OOB write
due to under allocation of the interpreter stack frame.

This CL adds a SBXCHECK for the wrapped argument counts against
Code::kMaxArguments during wrapped arguments parsing.

TAG=agy

Fixed: 554034656
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/8368411
Commit-Queue: Arash Kazemi <arashk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#109772}

+++ src/interpreter/bytecode-array-builder.cc
@@ -53,7 +53,7 @@ BytecodeArrayBuilder::BytecodeArrayBuilder(
       bytecode_generated_(false),
       constant_array_builder_(zone),
       handler_table_builder_(zone),
-      parameter_count_(parameter_count),
+      parameter_count_(base::checked_cast<uint16_t>(parameter_count)),
       max_arguments_(0),
       local_register_count_(locals_count),
       register_allocator_(fixed_register_count()),
